Rename layoutContext.tsx to chartLayoutContext.tsx
